maj len_nbr and strdup again

+++ parser/utils.c
@@ -6,7 +6,7 @@
 /*   By: bkhatib <bkhatib@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/22 22:10:01 by bkhatib           #+#    #+#             */
-/*   Updated: 2022/05/30 22:06:47 by bkhatib          ###   ########.fr       */
+/*   Updated: 2022/05/30 22:51:12 by bkhatib          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,8 +45,8 @@ size_t	ft_strlen(const char *str)
 char	*ft_strdup(const char *str)
 {
 	char	*dup;
-	int	str_len;
-	int	i;
+	size_t	str_len;
+	size_t	i;
 	
 	i = 0;
 	str_len = ft_strlen(str);
@@ -133,7 +133,7 @@ int	in_set(char c)
 	return (1);
 }
 
-int	len_nbr(int n)
+size_t	len_nbr(int n)
 {
 	size_t	nbr;
